refactor: refactor position of types.tsx & fix error
