+++ products_application/src/types/types.tsx
